feat: add client package which removes base modules exposure; remove redundant/refactored files & update tests
