+++ src/client.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	"github.com/extended-protocol/extended-sdk-golang/src/account"
+	"github.com/extended-protocol/extended-sdk-golang/src/client"
 	"github.com/extended-protocol/extended-sdk-golang/src/markets"
 	"github.com/extended-protocol/extended-sdk-golang/src/models"
 	"github.com/extended-protocol/extended-sdk-golang/src/orders"
@@ -13,7 +14,7 @@ import (
 // It embeds BaseModule to reuse common functionality like HTTP client, auth, etc.
 // It provides access to domain-specific services through the Account, Orders, and Markets fields.
 type Client struct {
-	*BaseModule
+	*client.BaseModule
 	Account *account.Service
 	Orders  *orders.Service
 	Markets *markets.Service
@@ -23,20 +24,20 @@ type Client struct {
 // It takes an endpoint configuration, a Stark perpetual account, and a client timeout.
 func NewClient(
 	cfg models.EndpointConfig,
-	starkAccount *StarkPerpetualAccount,
+	starkAccount *client.StarkPerpetualAccount,
 	clientTimeout time.Duration,
 ) *Client {
-	baseModule := NewBaseModule(cfg, starkAccount.APIKey(), starkAccount, nil, clientTimeout)
-	client := &Client{
+	baseModule := client.NewBaseModule(cfg, starkAccount.APIKey(), starkAccount, nil, clientTimeout)
+	sdkClient := &Client{
 		BaseModule: baseModule,
 	}
 
-	// Initialize services with reference to the client
-	client.Account = &account.Service{Client: client}
-	client.Orders = &orders.Service{Client: client}
-	client.Markets = &markets.Service{Client: client}
+	// Initialize services with reference to BaseModule
+	sdkClient.Account = &account.Service{Base: baseModule}
+	sdkClient.Orders = &orders.Service{Base: baseModule}
+	sdkClient.Markets = &markets.Service{Base: baseModule}
 
-	return client
+	return sdkClient
 }
 
 // Close closes the HTTP client and cleans up resources.
