+++ src/client/account.go
@@ -1,0 +1,78 @@
+package client
+
+import (
+	"errors"
+	"fmt"
+	"math/big"
+)
+
+// StarkPerpetualAccount represents a Stark perpetual trading account.
+type StarkPerpetualAccount struct {
+	vault      uint64
+	privateKey string
+	publicKey  string
+	apiKey     string
+}
+
+// NewStarkPerpetualAccount constructs the account, validating hex inputs.
+func NewStarkPerpetualAccount(vault uint64, privateKeyHex, publicKeyHex, apiKey string) (*StarkPerpetualAccount, error) {
+	if err := isHexString(privateKeyHex); err != nil {
+		return nil, fmt.Errorf("invalid private key: %w", err)
+	}
+	if err := isHexString(publicKeyHex); err != nil {
+		return nil, fmt.Errorf("invalid public key: %w", err)
+	}
+
+	// Ensure that private key and public key have 0x prefix
+	if len(privateKeyHex) < 2 || privateKeyHex[:2] != "0x" {
+		return nil, fmt.Errorf("private key must start with 0x")
+	}
+	if len(publicKeyHex) < 2 || publicKeyHex[:2] != "0x" {
+		return nil, fmt.Errorf("public key must start with 0x")
+	}
+
+	// Check that API key does not start with 0x
+	if len(apiKey) >= 2 && apiKey[:2] == "0x" {
+		return nil, fmt.Errorf("api key should not start with 0x")
+	}
+
+	return &StarkPerpetualAccount{
+		vault:      vault,
+		privateKey: privateKeyHex,
+		publicKey:  publicKeyHex,
+		apiKey:     apiKey,
+	}, nil
+}
+
+// Vault returns the vault id.
+func (s *StarkPerpetualAccount) Vault() uint64 { return s.vault }
+
+// PublicKey returns the public key as a string.
+func (s *StarkPerpetualAccount) PublicKey() string { return s.publicKey }
+
+// APIKey returns the API key string.
+func (s *StarkPerpetualAccount) APIKey() string { return s.apiKey }
+
+// Sign delegates to SignFunc, returning (r,s).
+func (stark *StarkPerpetualAccount) Sign(msgHash string) (*big.Int, *big.Int, error) {
+	if msgHash == "" {
+		return big.NewInt(0), big.NewInt(0), errors.New("msgHash is empty")
+	}
+
+	sig, err := SignMessage(msgHash, stark.privateKey)
+	if err != nil {
+		return big.NewInt(0), big.NewInt(0), err
+	}
+
+	// Extract r, s from the signature string.
+	// Signature is in the format of {r}{s}{v}, where r, s and v are 64 chars each (192 hex chars).
+	r, isGoodR := big.NewInt(0).SetString(sig[:64], 16)
+	s, isGoodS := big.NewInt(0).SetString(sig[64:128], 16)
+
+	if !isGoodR || !isGoodS {
+		return big.NewInt(0), big.NewInt(0), errors.New("big int setting failed")
+	}
+
+	return r, s, nil
+}
+
