+++ src/client/sign.go
@@ -1,8 +1,8 @@
-package sdk
+package client
 
 /*
-#cgo LDFLAGS: -L${SRCDIR}/../ -lorderffi
-#cgo LDFLAGS: -Wl,-rpath,${SRCDIR}/../
+#cgo LDFLAGS: -L${SRCDIR}/../../../ -lorderffi
+#cgo LDFLAGS: -Wl,-rpath,${SRCDIR}/../../../
 #include <stdlib.h>
 
 char* get_order_hash_ffi(
@@ -93,3 +93,4 @@ func SignMessage(messageHex, privateKeyHex string) (string, error) {
 
 	return C.GoString(sig), nil
 }
+
